% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/status.R
\name{att_status}
\alias{att_status}
\title{View Source Status Summary}
\usage{
att_status(store = NULL)
}
\arguments{
\item{store}{Path to the store. Defaults to \code{\link[=att_store]{att_store()}}.}
}
\value{
A data frame of source information.
}
\description{
Prints a summary table of all tracked sources in the store by scanning
for \verb{_acquire/provenance.json} files.
}
\examples{
\dontrun{
att_status()
}
}
