% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/attest-package.R
\docType{package}
\name{attest-package}
\alias{attest}
\alias{attest-package}
\title{attest: fail-closed machine learning}
\description{
Models that carry their own proof of validity. See \code{\link[=attest_spec]{attest_spec()}},
\code{\link[=attest_fit]{attest_fit()}}, \code{\link[=predict.attested_model]{predict.attested_model()}} and \code{\link[=report]{report()}}.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/diogoribeiro7/attest}
  \item Report bugs at \url{https://github.com/diogoribeiro7/attest/issues}
}

}
\author{
\strong{Maintainer}: Diogo Ribeiro \email{dfr@esmad.ipp.pt} (\href{https://orcid.org/0009-0001-2022-7072}{ORCID})

}
\keyword{internal}
