% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/engines.R
\name{engine_fit}
\alias{engine_fit}
\title{Fit a model with an engine}
\usage{
engine_fit(engine, formula, data, task)
}
\arguments{
\item{engine}{An \code{attest_engine}.}

\item{formula}{Model formula.}

\item{data}{Training data.}

\item{task}{\code{"classification"} or \code{"regression"}.}
}
\value{
A fitted object.
}
\description{
Fit a model with an engine
}
