% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/engines.R
\name{engines}
\alias{engines}
\alias{engine_glm}
\alias{engine_ranger}
\title{Engine adapters}
\usage{
engine_glm(...)

engine_ranger(...)
}
\arguments{
\item{...}{Arguments passed to the underlying fitting function.}
}
\value{
An object of class \code{attest_engine}.
}
\description{
An engine is a thin adapter around a modelling function. Three generics
define the interface: \code{\link[=engine_fit]{engine_fit()}} and \code{\link[=engine_predict]{engine_predict()}}.
}
